Revert some of it.
