+++ measures/HPXMLtoOpenStudio/measure.rb
@@ -4137,5 +4137,86 @@ def get_fan_power_installed(seer)
   end
 end
 
+class OutputVars
+  def self.SpaceHeatingElectricity
+    return { 'OpenStudio::Model::CoilHeatingDXSingleSpeed' => ['Heating Coil Electric Energy', 'Heating Coil Crankcase Heater Electric Energy', 'Heating Coil Defrost Electric Energy'],
+             'OpenStudio::Model::CoilHeatingDXMultiSpeed' => ['Heating Coil Electric Energy', 'Heating Coil Crankcase Heater Electric Energy', 'Heating Coil Defrost Electric Energy'],
+             'OpenStudio::Model::CoilHeatingElectric' => ['Heating Coil Electric Energy', 'Heating Coil Crankcase Heater Electric Energy', 'Heating Coil Defrost Electric Energy'],
+             'OpenStudio::Model::CoilHeatingWaterToAirHeatPumpEquationFit' => ['Heating Coil Electric Energy', 'Heating Coil Crankcase Heater Electric Energy', 'Heating Coil Defrost Electric Energy'],
+             'OpenStudio::Model::CoilHeatingGas' => [],
+             'OpenStudio::Model::ZoneHVACBaseboardConvectiveElectric' => ['Baseboard Electric Energy'],
+             'OpenStudio::Model::BoilerHotWater' => ['Boiler Electric Energy'],
+             'OpenStudio::Model::FanOnOff' => ['Fan Electric Energy'] }
+  end
+
+  def self.SpaceHeatingFuel
+    return { 'OpenStudio::Model::CoilHeatingDXSingleSpeed' => [],
+             'OpenStudio::Model::CoilHeatingDXMultiSpeed' => [],
+             'OpenStudio::Model::CoilHeatingElectric' => [],
+             'OpenStudio::Model::CoilHeatingWaterToAirHeatPumpEquationFit' => [],
+             'OpenStudio::Model::CoilHeatingGas' => ['Heating Coil Gas Energy', 'Heating Coil Propane Energy', 'Heating Coil FuelOil#1 Energy'],
+             'OpenStudio::Model::ZoneHVACBaseboardConvectiveElectric' => ['Baseboard Gas Energy', 'Baseboard Propane Energy', 'Baseboard FuelOil#1 Energy'],
+             'OpenStudio::Model::BoilerHotWater' => ['Boiler Gas Energy', 'Boiler Propane Energy', 'Boiler FuelOil#1 Energy'],
+             'OpenStudio::Model::FanOnOff' => [] }
+  end
+
+  def self.SpaceHeatingLoad
+    return { 'OpenStudio::Model::CoilHeatingDXSingleSpeed' => ['Heating Coil Heating Energy'],
+             'OpenStudio::Model::CoilHeatingDXMultiSpeed' => ['Heating Coil Heating Energy'],
+             'OpenStudio::Model::CoilHeatingElectric' => ['Heating Coil Heating Energy'],
+             'OpenStudio::Model::CoilHeatingWaterToAirHeatPumpEquationFit' => ['Heating Coil Heating Energy'],
+             'OpenStudio::Model::CoilHeatingGas' => ['Heating Coil Heating Energy'],
+             'OpenStudio::Model::ZoneHVACBaseboardConvectiveElectric' => ['Baseboard Total Heating Energy'],
+             'OpenStudio::Model::BoilerHotWater' => ['Boiler Heating Energy'],
+             'OpenStudio::Model::FanOnOff' => ['Fan Electric Energy'] }
+  end
+
+  def self.SpaceCoolingElectricity
+    return { 'OpenStudio::Model::CoilCoolingDXSingleSpeed' => ['Cooling Coil Electric Energy', 'Cooling Coil Crankcase Heater Electric Energy'],
+             'OpenStudio::Model::CoilCoolingDXMultiSpeed' => ['Cooling Coil Electric Energy', 'Cooling Coil Crankcase Heater Electric Energy'],
+             'OpenStudio::Model::CoilCoolingWaterToAirHeatPumpEquationFit' => ['Cooling Coil Electric Energy', 'Cooling Coil Crankcase Heater Electric Energy'],
+             'OpenStudio::Model::FanOnOff' => ['Fan Electric Energy'] }
+  end
+
+  def self.SpaceCoolingLoad
+    return { 'OpenStudio::Model::CoilCoolingDXSingleSpeed' => ['Cooling Coil Total Cooling Energy'],
+             'OpenStudio::Model::CoilCoolingDXMultiSpeed' => ['Cooling Coil Total Cooling Energy'],
+             'OpenStudio::Model::CoilCoolingWaterToAirHeatPumpEquationFit' => ['Cooling Coil Total Cooling Energy'],
+             'OpenStudio::Model::FanOnOff' => ['Fan Electric Energy'] }
+  end
+
+  def self.WaterHeatingElectricity
+    return { 'OpenStudio::Model::WaterHeaterMixed' => ['Water Heater Electric Energy', 'Water Heater Off Cycle Parasitic Electric Energy', 'Water Heater On Cycle Parasitic Electric Energy'],
+             'OpenStudio::Model::WaterHeaterStratified' => ['Water Heater Electric Energy', 'Water Heater Off Cycle Parasitic Electric Energy', 'Water Heater On Cycle Parasitic Electric Energy'],
+             'OpenStudio::Model::CoilWaterHeatingAirToWaterHeatPumpWrapped' => ['Cooling Coil Water Heating Electric Energy'],
+             'OpenStudio::Model::WaterUseConnections' => [],
+             'OpenStudio::Model::ElectricEquipment' => [] }
+  end
+
+  def self.WaterHeatingElectricityRecircPump
+    return { 'OpenStudio::Model::WaterHeaterMixed' => [],
+             'OpenStudio::Model::WaterHeaterStratified' => [],
+             'OpenStudio::Model::CoilWaterHeatingAirToWaterHeatPumpWrapped' => [],
+             'OpenStudio::Model::WaterUseConnections' => [],
+             'OpenStudio::Model::ElectricEquipment' => ['Electric Equipment Electric Energy'] }
+  end
+
+  def self.WaterHeatingFuel
+    return { 'OpenStudio::Model::WaterHeaterMixed' => ['Water Heater Gas Energy', 'Water Heater Propane Energy', 'Water Heater FuelOil#1 Energy'],
+             'OpenStudio::Model::WaterHeaterStratified' => ['Water Heater Gas Energy', 'Water Heater Propane Energy', 'Water Heater FuelOil#1 Energy'],
+             'OpenStudio::Model::CoilWaterHeatingAirToWaterHeatPumpWrapped' => [],
+             'OpenStudio::Model::WaterUseConnections' => [],
+             'OpenStudio::Model::ElectricEquipment' => [] }
+  end
+
+  def self.WaterHeatingLoad
+    return { 'OpenStudio::Model::WaterHeaterMixed' => [],
+             'OpenStudio::Model::WaterHeaterStratified' => [],
+             'OpenStudio::Model::CoilWaterHeatingAirToWaterHeatPumpWrapped' => [],
+             'OpenStudio::Model::WaterUseConnections' => ['Water Use Connections Plant Hot Water Energy'],
+             'OpenStudio::Model::ElectricEquipment' => [] }
+  end
+end
+
 # register the measure to be used by the application
 HPXMLTranslator.new.registerWithApplication
